+++ R/html.R
@@ -71,8 +71,10 @@ to_html.huxtable <- function(ht, ...) {
   lab <- make_label(ht)
   id_string <- if (is.na(lab)) "" else sprintf(" id=\"%s\"", lab)
 
-  table_start <- sprintf(
-        '<table class="huxtable" style="border-collapse: collapse; border: 0px; margin-bottom: 2em; margin-top: 2em; %s; %s %s %s"%s>\n',
+  table_start <- sprintf(paste0(
+        '<table class="huxtable" quarto-disable-processing="true" ',
+        'style="border-collapse: collapse; border: 0px; ',
+        'margin-bottom: 2em; margin-top: 2em; %s; %s %s %s"%s>\n'),
         width_string, margin_string, height_string, float_string, id_string)
 
   if (! is.na(cap <- make_caption(ht, lab, "html"))) {