+++ NEWS.md
@@ -4,6 +4,7 @@
   `.` is now vector rather than matrix when dplyr version 1.1.0 is detected. 
   Thanks @DavisVaughan.
 * Add package checks in `quick_*` functions. Thanks @reuning.
+* Bugfix: disable quarto styling on HTML tables.
 
 # huxtable 5.5.1